add copyToClipboard tests and fix related tests.
